backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-JWT-Extended==4.4.4
PyJWT==2.8.0
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
